% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl.R
\name{afl}
\alias{afl}
\title{(Alias for) \code{add_frame_lab()}}
\usage{
afl(data, frame.lab = NULL)
}
\arguments{
\item{data}{a data.frame.}

\item{frame.lab}{quoted text of the descriptive data.frame label that you wish
to add to the data.frame.}
}
\value{
A data.frame, with a frame.lab attribute added to the attributes
meta-data
}
\description{
\code{afl()} is an alias for \code{add_frame_lab()}.
}
\details{
\code{afl} assigns an overall descriptive "frame label" for a data.frame,
which can be retrieved using \code{get_frame_lab} or its alias \code{gfl}.
}
\examples{
# add frame.lab to mtcars and assign to new data.frame mt2
mt2 <- afl(mtcars, frame.lab = "Data extracted from the 1974 Motor
Trend US magazine, comprising fuel consumption and 10 aspects of automobile
design and performance for 32 automobiles (1973–74 models). Source: Henderson
and Velleman (1981), Building multiple regression models interactively.
 Biometrics, 37, 391–411.")
attr(mt2, "frame.lab") # check for attribute
gfl(mt2) # return frame.lab alongside data.frame name as a data.frame
}
