% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aql1.R
\name{aql1}
\alias{aql1}
\title{(Alias for) \code{add_quant1()}}
\usage{
aql1(data, var, qtiles = NULL, vals = NULL, labs = NULL)
}
\arguments{
\item{data}{a data.frame.}

\item{var}{the unquoted name of the variable to which value labels will be
added.}

\item{qtiles}{the number of quantile categories to employ (e.g., 4 would
indicate quartiles, 5 would indicate quintiles, 10 for deciles, etc.). If
NULL, vals must be non-NULL.}

\item{vals}{one more values of var that will define range cutpoints, such
that all values at or below a given number and above the preceding val will
be treated as part of the same numerical range for labeling purposes. If
NULL, qtiles must be non-NULL.}

\item{labs}{a character vector of distinct labels to identify the
quantiles. If left NULL, convention "q" + quantile (e.g., "q10") will be used
for qtile-based labels (i.e., if qtiles arg is non-NULL), and convention
"<=" + val will be used for vals argument-based labels (i.e., if vals arg is
non-NULL). Note that the labels "NA" and "Other" are (non-case-sensitively)
reserved and may not be user-supplied.}
}
\value{
A data.frame, with new variable value labels added (call
\code{get_val_labs} to see them), other provisional/default labelr label
information added, and previous user-added labelr label information
preserved.
}
\description{
Add variable-specific value labels based on threshold cuts of a single
numerical variable.
}
\details{
Note: \code{aql1} is an alias for \code{add_quant1}. See that command's documentation for
further information.
}
\examples{
# mtcars demo
df <- mtcars
# now, add value labels
df <- add_val_labs(
  data = df,
  vars = "am",
  vals = c(0, 1),
  labs = c("automatic", "manual")
)

# label variable "mpg" in terms of 5 quintiles
df <- aql1(data = df, mpg, qtiles = 5)

# label variable "disp" in terms of "pretty" cutpoints
vals2use <- pretty(c(min(df$disp), max(df$disp)))[-1] # establish cutpoints
df <- aql1(data = df, disp, vals = vals2use)
df_labson <- use_val_labs(df)
head(df_labson)
}
