% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditWorkers.R
\name{auditWorkers}
\alias{auditWorkers}
\title{Audit Mechanical Turk workers}
\usage{
auditWorkers(current_experiment_results, reference_results = NULL)
}
\arguments{
\item{current_experiment_results}{An object returned from \code{getResults}
containing HIT results from Mechanical Turk.}

\item{reference_results}{An optional object of HIT results from previous
experiments. When available, this improves the accuracy of worker quality scores.}
}
\value{
A dataframe containing workers' posterior means and HIT frequencies.
}
\description{
\code{auditWorkers} estimates worker quality scores from collected MTurk data.
}
\details{
The function audits worker quality using a Bayesian hierarchical model. Worker
means less than one are generally considered poor quality, though means can be
unreliable for workers with a low number of total HITs.
}
\author{
Ryden Butler
}
