% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnl.R
\name{dnl}
\alias{dnl}
\title{(Alias for) \code{drop_name_labs()}}
\usage{
dnl(data, vars = NULL)
}
\arguments{
\item{data}{the data.frame with name labels that you wish to drop (and which
were added using \code{add_name_labs}).}

\item{vars}{the names of the columns/variables (not the name labels) whose
name labels you wish to drop. If NULL, all variables' name labels will be
dropped.}
}
\value{
The same data.frame you submitted, except that the selected name
label attribute meta-data has been removed.
}
\description{
Remove one or more descriptive variable name label attributes previously
added to a data.frame using \code{add_name_labs}.
}
\details{
\code{dnl} is an alias for \code{drop_name_labs()}.
}
\examples{
# create a data set
df <- mtcars

# variable names and their labels
names_labs_vec <- c(
  "mpg" = "Miles/(US) gallon",
  "cyl" = "Number of cylinders",
  "disp" = "Displacement (cu.in.)",
  "hp" = "Gross horsepower",
  "drat" = "Rear axle ratio",
  "wt" = "Weight (1000 lbs)",
  "qsec" = "1/4 mile time",
  "vs" = "Engine (0 = V-shaped, 1 = straight)",
  "am" = "Transmission (0 = automatic, 1 = manual)",
  "gear" = "Number of forward gears",
  "carb" = "Number of carburetors"
)

# assign variable labels
df <- add_name_labs(df,
  vars = names(names_labs_vec),
  labs = names_labs_vec
)

# see what we have
get_name_labs(df)

# drop the name label for var/col "am"
df <- dnl(df, "am")

# see what this did to the name label for "am"
get_name_labs(df)

# now, drop all of the name labels
df <- dnl(df)
get_name_labs(df) # they're gone
}
