% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact2char.R
\name{fact2char}
\alias{fact2char}
\alias{f2c}
\title{Convert All Factor Variables of a Data Frame to Column Variables}
\usage{
fact2char(data)

f2c(data)
}
\arguments{
\item{data}{a data.frame object.}
}
\value{
a data.frame identical to data, with exception that any factors
have been converted to character variables.
}
\description{
Convenience function to convert all factor variables to character.
}
\examples{
sapply(iris, class)
head(iris)

iris_ch <- fact2char(iris)

sapply(iris_ch, class)
head(iris_ch)
}
