% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_lab_int.R
\name{factor_to_lab_int}
\alias{factor_to_lab_int}
\alias{f2int}
\title{Convert a Factor Variable Column to Value-labeled Integer Variable Column}
\usage{
factor_to_lab_int(data, var)

f2int(data, var)
}
\arguments{
\item{data}{a data.frame object.}

\item{var}{the (unquoted) name of a factor variable found in data.}
}
\value{
a data.frame.
}
\description{
\code{factor_to_lab_int} converts a factor variable (column) of a data.frame to a
value-labeled integer variable and converts the factor level labels to labelr
value labels, returning the modified data.frame.
}
\details{
Note: \code{f2int} is a compact alias for \code{factor_to_lab_int}: they do the same thing,
and the former is easier to type.
}
\examples{
class(iris[["Species"]])
iris_df <- factor_to_lab_int(iris, Species)
class(iris_df[["Species"]])
get_val_labs(iris_df, "Species")
}
