% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitStan.R
\name{fitStan}
\alias{fitStan}
\title{Fit a Stan model with results from MTurk}
\usage{
fitStan(data, chains = 3, iter = 2500, seed = 1234, n.cores = 3)
}
\arguments{
\item{data}{A data set or a vector of batch numbers.}

\item{chains}{The number of chains. (Default is 3)}

\item{iter}{The number of iteration. (Default is 2500)}

\item{seed}{Set seed. (Defalt is 1234)}

\item{n.cores}{Number of cores to be used in stan fit. (Default is 3)}
}
\value{
fit A list containing the following elements:
\itemize{
\item fit The Stan fit object for the model
\item alphaPosts A matrix with the full posteriors of the document estimates merged with the document IDs from the SentimentIt server
}
}
\description{
Fit a Stan model with results from MTurk
}
\details{
\code{fitStan()} function estimates a non-hierarchical Stan model,

Fit a random utility model using Hamiltonian MCMC in Stan with data retrieved from the SentimentIt platform.

Reference Paper: Carlson, David and Jacob M. Montgomery. Forthcoming. “A Pairwise Comparison Framework for
Fast, Flexible, and Reliable Human Coding of Political Texts.” American Political Science Review.
}
\examples{
\dontrun{
data(movieReviewOutput)
fit <- fitStan(data = movieReviewOutput) # can alternatively be batch IDs
fit <- fitStan(data = output)
}

}
\author{
David Carlson
}
