% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatAnswer.R
\name{formatAnswers}
\alias{formatAnswers}
\title{Format a MTurk qualification test answer key}
\usage{
formatAnswers(question_list, question_number)
}
\arguments{
\item{question_list}{A list object where the first and second element
correspond to the left and right comparison text, respectively, and the third
element corresponds to the index of the correct choice.}

\item{question_number}{An integer referencing the question number as it
appears in the MTurk qualification test.}
}
\value{
A string of XML code that constructs the answer key for an
MTurk qualification test.
}
\description{
\code{formatAnswer} formats the XML code required to make an answer key for
an MTurk qualification test.
}
\details{
The function automatically formats the XML code required to make a generic
answer key for an MTurk qualification test.
}
\author{
Ryden Butler
}
