% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatQuestion.R
\name{formatQuestion}
\alias{formatQuestion}
\title{Format a MTurk qualification test question}
\usage{
formatQuestion(
  question_list,
  question_number,
  question_prompt,
  is_image = F,
  test_question = F
)
}
\arguments{
\item{question_list}{A list object where the first and second element
contain the left and right comparison text, respectively, and the third
element contains an integer corresponding to the index of the correct choice.}

\item{question_number}{An integer referencing the question number as it
appears in the MTurk qualification test.}

\item{question_prompt}{A string containing the common question prompt used
across all pairwise comparison questions.}

\item{is_image}{A logical where TRUE will cause the first two
\code{question_list} elements to be treated as images for formatting purposes.
FALSE is default and treats the first two \code{question_list} elements as html-formatted text.}

\item{test_question}{A logical where FALSE indicates that the question is
a practice question and TRUE indicates that the question is a test question.}
}
\value{
A string of XML code that constructs a question for an
MTurk qualification test.
}
\description{
\code{formatQuestion} formats the XML code required to make a pairwise comparison
question for a MTurk qualification test.
}
\details{
The function automatically formats the XML code required to make a generic
pairwise comparison question for an MTurk qualification test.
}
\author{
Ryden Butler
}
