% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTest.R
\name{formatTest}
\alias{formatTest}
\title{Format a MTurk qualification test}
\usage{
formatTest(title, instruction_overview, instruction_list, question_prompt,
  practice_overview, practice_questions, test_overview, test_questions,
  closing_message)
}
\arguments{
\item{title}{A string containing the title of the qualification test, as it
appears within the test.}

\item{instruction_overview}{A string providing general instructions for the HITs}

\item{instruction_list}{A nested list of strings containing bulleted instructions
and examples. Each element of the list object corresponds to a separate bullet
in the list of instructions. Nesting list objects within one another will allow
for nested bullets.}

\item{question_prompt}{A string containing the common question prompt used
across all pairwise comparison questions.}

\item{practice_overview}{A string providing general instructions for the
practice questions within the qualification test. This should inform workers
how practice questions differ from other questions in the test and/or HITs
on MTurk.}

\item{practice_questions}{A list of lists where each element of the main list
contains a single list for each unique practice question. Each question-specific
list contains three elements, where the first and second element
contain the left and right comparison text, respectively, and the third
element contains text explaining the correct answer.}

\item{test_overview}{A string providing general instructions for the
test questions within the qualification test. This should contain information
about scoring procedures.}

\item{test_questions}{A list of lists where each element of the main list
contains a single list for each unique practice question. Each question-specific
list contains three elements, where the first and second element
contain the left and right comparison text, respectively, and the third
element contains an integer corresponding to the index of the correct choice.}

\item{closing_message}{A string containing the final text presented in the
training module. This should contain warnings about worker monitoring and/or
rate limiting procedures.}
}
\value{
A list of XML formatted strings that specify the qualification test
and answer key. Each is written as a .xml file in the working directory for
users to make manual revisions.
}
\description{
\code{formatTest} formats the XML code required to make a MTurk qualification
test.
}
\details{
The function writes generic .xml files required to make an MTurk qualification
test. Though the function is designed for users to pass strings directly into
the arguments, the arguments also accept HTML-formatted text.
}
\author{
Ryden Butler
}
