% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_lab_atts.R
\name{get_all_lab_atts}
\alias{get_all_lab_atts}
\title{Put all Data Frame label attributes into a List}
\usage{
get_all_lab_atts(data, atts = c("name.labs|val.labs|frame.lab|factor."))
}
\arguments{
\item{data}{the data.frame you are checking for labelr meta-data attributes.}

\item{atts}{default is to look for all/any, but you can specify a more
narrow subset (or some other, altogether-irrelevant attribute, but do not do
that).}
}
\value{
A free-standing list of labelr attributes.
}
\description{
\code{get_all_lab_atts} returns a list of labelr-generated meta-data attributes
attached to a data.frame, all of which should have names beginning with one
of these character strings: "frame.lab", "name.labs", \cr
"val.labs", "factor.".
}
\examples{
# make toy demographic (gender, raceth, etc.) data set
df <- make_demo_data(n = 1000, seed = 555) # another labelr:: function
# let's add variable VALUE labels for variable "raceth"
df <- add_val_labs(df,
  vars = "raceth", vals = c(1:7),
  labs = c("White", "Black", "Hispanic", "Asian", "AIAN", "Multi", "Other"),
  max.unique.vals = 50
)

get_all_lab_atts(df) # returns all; is default
get_all_lab_atts(df, "val.labs") # returns only "val.labs" attributes
get_all_lab_atts(df, "class") # You can (but probably should not) use this way.
}
