% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Get results from Mechanical Turk}
\usage{
getResults(
  current_hit_ids,
  current_document_ids,
  current_batch_id = 0,
  retry = T,
  retry_in_seconds = 60,
  hit_categories = c("Document 1", "Document 2")
)
}
\arguments{
\item{current_hit_ids}{A vector of HIT ids to be retieved.}

\item{current_document_ids}{A 2-column matrix of document ids corresponding
to the documents included in each HIT.}

\item{current_batch_id}{An integer used to associate the current results with
a desired batch}

\item{retry}{A logical where TRUE causes the function to repeat until all
results are downloaded, providing live progress updates throughout. When FALSE,
the function will only download all currently available results for the specified
HITs.}

\item{retry_in_seconds}{An integer specifying how long the function should pause
before retrying to collect results. Only applicable when \code{retry = T}.}

\item{hit_categories}{A vector of strings with crowd classifier category names
as they appear in the HITformat.html file. These should not be modified unless
using a custom version of HITformat.html where alternate category names have
been specified.}
}
\value{
A list containing a data.frame of results from MTurk and a vector of 
HIT ids of incomplete HITs.
}
\description{
\code{getResults} downloads completed work from MTurk for the specified HITs
and can offer live progress updates on the amount of completed work.
}
\details{
The function retrieves and formats labeled data from MTurk.
}
\author{
Ryden Butler
}
