% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfl.R
\name{gfl}
\alias{gfl}
\title{(Alias for) \code{get_frame_lab()}}
\usage{
gfl(data)
}
\arguments{
\item{data}{a data.frame.}
}
\value{
A 1x2 data.frame, consisting of "data.frame" and "frame.lab" columns
for supplied data.frame.
}
\description{
\code{gfl()} is an alias for \code{get_frame_lab()}.
}
\details{
\code{gfl} returns the overall descriptive "frame label" that is assigned
to a data.frame using \code{add_frame_lab} or its alias \code{afl}.
}
\examples{
# add frame.lab to mtcars and assign to new data.frame mt2
mt2 <- add_frame_lab(mtcars, frame.lab = "Data extracted from the 1974 Motor
                      Trend US magazine, comprising fuel consumption and 10
                      aspects of automobile design and performance for 32
                      automobiles (1973–74 models). Source: Henderson and
                      Velleman (1981), Building multiple regression models
                      interactively. Biometrics, 37, 391–411.")

attr(mt2, "frame.lab") # check for attribute

# return frame.lab alongside data.frame name as a data.frame
gfl(mt2)
}
