% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_int_to_factor.R
\name{lab_int_to_factor}
\alias{lab_int_to_factor}
\alias{int2f}
\title{Convert a Value-labeled Integer Variable Column to a Factor Variable Column}
\usage{
lab_int_to_factor(data, var)

int2f(data, var)
}
\arguments{
\item{data}{a data.frame object.}

\item{var}{the (unquoted) name of a value-labeled integer variable found in
data.}
}
\value{
a data.frame.
}
\description{
\code{lab_int_to_factor} converts a value-labeled integer variable to a factor,
using labelr value labels as factor level labels and returning the modified
data.frame.
}
\details{
Note: \code{int2f} is a compact alias for \code{lab_int_to_factor}: they do the same
thing, and the former is easier to type.
}
\examples{
class(iris[["Species"]])

iris_sp_int <- factor_to_lab_int(iris, Species)
class(iris_sp_int[["Species"]])

get_val_labs(iris_sp_int, "Species")
iris_sp_fac <- lab_int_to_factor(iris_sp_int, Species)

class(iris_sp_fac[["Species"]])

levels(iris_sp_fac[["Species"]])
}
