% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_df.R
\name{label_df}
\alias{label_df}
\title{Label data frame}
\usage{
label_df(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Pass further arguments to \code{label_var()}.}
}
\value{
Labelled data frame
}
\description{
Use this to label your final data frame before you plot it.
}
\examples{
label_df(df)
}
\author{
Tino Good, \email{onit.good@gmail.com}
}
