% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_var.R
\name{label_var}
\alias{label_var}
\title{Label variable values}
\usage{
label_var(x, variable = NULL, ...)
}
\arguments{
\item{x}{Vector}

\item{variable}{Denote which variable vector represents. If not changed from defaults, it will find out.}

\item{...}{Other filters.}
}
\value{
Vector of labels
}
\description{
Label variable values
}
\author{
Tino Good, \email{onit.good+labelr@gmail.com}
}
