% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify.R
\name{listify}
\alias{listify}
\title{Format a HTML unordered list}
\usage{
listify(nested_list)
}
\arguments{
\item{nested_list}{A list object containing a combination of strings or lists.
Each string element of the list is reformatted as a bullet in an unordered list.
Each list object element is treated as a nested unordered list, which will possess
a different bullet type.}
}
\value{
An html-formatted list
}
\description{
\code{listify} reformats list objects as unordered HTML lists.
}
\details{
This function is for internal package use. It creates a string in HTML format
that renders an unordered list when provided with a \code{list} object.
}
\author{
Ryden Butler
}
