% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_defaults.R
\name{new_defaults}
\alias{new_defaults}
\title{New label defaults}
\usage{
new_defaults(value = list())
}
\arguments{
\item{value}{Named list of data frames with at least \code{from} and \code{to} values (columns).}
}
\description{
New label defaults
}
\author{
Yihui Xie (borrowed from knitr)
}
