% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\docType{data}
\name{Text}
\alias{Text}
\title{Example text data for creating a HIT and qualification test on MTurk}
\format{
A list with 10 elements.
\describe{
\item{Title}{String. The title of the qualification test.}
\item{Instruction Overview}{String. Detailed instructions on how workers are
to complete HITs.}
\item{InstructionList}{List of lists. This content will create a bulleted list
of key rules and simple examples to explain the task to workers. If an element
of the primary list is a string, it is converted to a bullet in the training
module. If an element of the primary list is a list object, the first element
of this sub-list should be a string, which will be converted to a bullet
in the training module. The second element in the sub-list should also be
a list containing at least one string-formatted element. The elements of this
list will be converted to sub-bullets in the training module, which are indented
with respect to the top-level bullets and will possess a different bullet style.}
\item{QuestionPrompt}{String. The generic question prompt that applies to all HITs.}
\item{PracticeOverview}{String. Detailed explanation of the practice
section of the qualification test. These questions are not scored.}
\item{PracticeQuestion}{List of lists. This content will create enumerated
practice questions with explanations of the answer. Each element of the primary
list should be a list object. Each sub-list should contain 3 elements. The
first element and second element are each strings, containing the text for
each document to be compared. The first element will be rendered as Document 1
and the second element will be rendered as Document 2. The third element is
a string containing explanatory text describing which document the worker should
have chosen and why that is the correct choice.}
\item{TestOverview}{String. Detailed explanation of the test
section of the qualification test. These questions are scored.}
\item{TestQuestions}{List of lists. This content will create enumerated
test questions. Each element of the primary list should be a list object.
Each sub-list should contain 3 elements. The first element and second element
are each strings, containing the text for each document to be compared. The
first element will be rendered as Document 1 and the second element will be
rendered as Document 2. The third element is a numeric integer indicating
which document should be chosen by the worker, with 1 indicating Document 1
and 2 indicating Document 2.}
\item{ClosingMessage}{String. The final text workers read before submitting
their qualification test for review.}
\item{ShortInstructions}{List of lists. This content will create a bulleted
list (of dashes) containing simplified instruction that workers can view in
the sidebar of a HIT. Each element of the primary list is a sub-list. The
first element of the sub-list is a string that will be rendered as a sub-title
(a HTML h3 tag) in the list. This can be left blank using empty quotes. Subsequent
elements of the sub-list are strings, each of which will receive a separate
bullet (dash).
}
...
}
}
\usage{
data(Text)
}
\description{
A list of text objects used by labelR to automatically format the XML files
required to make an MTurk qualification test and the HTML file used to make
a HIT.
}
\keyword{datasets}
