% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbl.R
\name{wbl}
\alias{wbl}
\title{(Alias for) \code{with_both_labs()}}
\usage{
wbl(data, ...)
}
\arguments{
\item{data}{a data.frame with variable name labels and variable value labels.}

\item{...}{an additional expression passed to dots (quotes and dollar signs
are not needed or permitted).}
}
\description{
\code{wbl} instructs R function calls to swap in variable name labels
for column names AND variable value labels for variable values in the objects
they return or side effects they produce.
}
\details{
\code{wbl} is an alias for \code{with_both_labs()}. See the latter's documentation for
additional details.
}
\examples{
# assign mtcars to new data.frame mt2
mt2 <- mtcars

# add name labs
mt2 <- add_name_labs(mt2,
  name.labs = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)


# add many-to-1 value labels
mt2 <- add_m1_lab(
  data = mt2,
  vars = "gear",
  vals = 4:5,
  lab = "4+"
)

# add many-to-1 value labels
mt2 <- add_val_labs(
  data = mt2,
  vars = "am",
  vals = c(0, 1),
  lab = c("auto", "man")
)
wbl(mt2, t.test(mpg ~ am))
wbl(mt2, lm(mpg ~ am))
wbl(mt2, xtabs(~gear))
xtabs(~ mt2$gear)
}
