% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnl.R
\name{wnl}
\alias{wnl}
\title{(Alias for) \code{with_name_labs()}}
\usage{
wnl(data, ...)
}
\arguments{
\item{data}{a data.frame with name labels.}

\item{...}{an additional expression passed to dots (quotes and dollar signs
are not needed or permitted).}
}
\description{
\code{wnl} instructs R function calls to swap in variable name labels
for column names in the objects they return or side effects they produce.
}
\details{
\code{wnl} is an alias for \code{with_name_labs()}. See the latter's documentation for
additional details.
}
\examples{
# assign mtcars to new data.frame mt2
mt2 <- mtcars

# add name labs
mt2 <- add_name_labs(mt2,
  name.labs = c(
    "mpg" = "Miles/(US) gallon",
    "cyl" = "Number of cylinders",
    "disp" = "Displacement (cu.in.)",
    "hp" = "Gross horsepower",
    "drat" = "Rear axle ratio",
    "wt" = "Weight (1000 lbs)",
    "qsec" = "1/4 mile time",
    "vs" = "Engine (0 = V-shaped, 1 = straight)",
    "am" = "Transmission (0 = automatic, 1 = manual)",
    "gear" = "Number of forward gears",
    "carb" = "Number of carburetors"
  )
)


wnl(mt2, t.test(mpg ~ am))
wnl(mt2, lm(mpg ~ am))
wnl(mt2, summary(mt2))
wnl(mt2, cor(mt2))
wnl(mt2, xtabs(~gear))
xtabs(~ mt2$gear)
wnl(mt2, cor(mpg, carb))
wnl(mt2, hist(mpg))
wnl(mt2, plot(mpg, carb))
wnl(mt2, head(gear))
wnl(mt2, summary(mt2))
}
