% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvl.R
\name{wvl}
\alias{wvl}
\title{(Alias for) \code{with_val_labs()}}
\usage{
wvl(data, ...)
}
\arguments{
\item{data}{a data.frame with value-labeled columns.}

\item{...}{additional arguments passed to dots, typically an expression
involving a function called on unquoted variable(s) (see examples).}
}
\description{
\code{wvl} wraps a data.frame in \code{use_val_labs} and wraps the resulting
data.frame in \code{base::with} in support of \code{base::with}-like non-standard
evaluation (see examples).
}
\details{
\code{wnl} is an alias for \code{with_name_labs()}. See the latter's documentation for
additional details.
}
\examples{
# make toy demographic (gender, raceth, etc.) data set
df <- make_demo_data(n = 1000, seed = 555) # another labelr:: function
# let's add variable VALUE labels for variable "raceth"
df <- add_val_labs(df,
  vars = "raceth", vals = c(1:7),
  labs = c("White", "Black", "Hispanic", "Asian", "AIAN", "Multi", "Other"),
  max.unique.vals = 50
)

# let's add variable VALUE labels for variable "gender"
# note that, if we are labeling a single variable, we can use add_val1()
# distinction between add_val1() and add_val_labs() will become more meaningful
# when we get to our Likert example
df <- add_val1(
  data = df, gender, vals = c(0, 1, 2),
  labs = c("M", "F", "O"), max.unique.vals = 50
)

# "wvl" - with()-like function that swaps value labels out for value values
# compare with(df, ...) to wvl(df,...)
with(df, table(gender, raceth)) # without labels

# the same data (note that presentation order changes d/t alphabetical ordering)
wvl(df, table(gender, raceth)) # with labels
with(use_val_labs(df), table(gender, raceth)) # above is shorthand for this

# just raceth
with(df, table(raceth)) # with
wvl(df, table(raceth)) # wvl

# another use case
with(df, unique(raceth)) # with
wvl(df, unique(raceth)) # wvl

# another
with(df, modelr::typical(raceth)) # numerical median!
wvl(df, modelr::typical(raceth)) # modal label (not the median!)
}
